Added Daily Stocks Scraping plus shuffled Files
File Shuffling.
Added Prototype folder. Old Prototype code is stored here for future reference. It's unlikely to needed. But it's better to keep it tucked aside instead of being sorry later.

Renamed Scraping and Processing notebooks.
Notebooks starting with 1_ are for scraping, 2_ are for processing

Will add 3_ for combining data and 4_ for modeling

Stocks added to the stock folder in compressed CSV zip format